Using doubles instead of floats

+++ Oscillator.cpp
@@ -6,7 +6,7 @@ Oscillator::Oscillator()
 {
 	_tables = new WaveTable*[TABLE_COUNT];
 	for (int i = 0; i < TABLE_COUNT; ++i) {
-		float hfreq = convertRanges((float)i, 0.0f, (float)TABLE_COUNT, (float)MIN_FREQUENCY, (float)MAX_FREQUENCY);
+		double hfreq = convertRanges((double)i, 0.0, (double)TABLE_COUNT, (double)MIN_FREQUENCY, (double)MAX_FREQUENCY);
 		_tables[i] = new WaveTable(wtSawtooth, hfreq);
 	}
 
@@ -25,14 +25,14 @@ Oscillator::~Oscillator()
 	delete [] _tables;
 }
 
-float Oscillator::getNextSample()
+double Oscillator::getNextSample()
 {
-	float ti = convertRanges(_frequencyValue, 0.0f, 1.0f, 0.0f, (float)TABLE_COUNT);
+	double ti = convertRanges(_frequencyValue, 0.0, 1.0, 0.0, (double)TABLE_COUNT);
 
 	int tableIndex1 = static_cast<int>(floor(ti));
 	int tableIndex2 = tableIndex1 + 1 >= TABLE_COUNT ? -1 : tableIndex1 + 1;
-	float tableWeight1 = ti - tableIndex1;
-	float tableWeight2 = tableWeight1 == -1 ? 0.0f : 1.0f - tableWeight1;
+	double tableWeight1 = ti - tableIndex1;
+	double tableWeight2 = tableWeight1 == -1 ? 0.0 : 1.0 - tableWeight1;
 	if (tableIndex2 == -1) {
 		tableIndex2 = 0;	// This will not affect anything. Only here to prevent reading out of bounds
 	}
@@ -42,22 +42,22 @@ float Oscillator::getNextSample()
 	/* Get next sample using interpolation */
 	int baseIndex = floor(_currentTableIndex);
 	int nextIndex = baseIndex + _tableIndexIncrement >= sampleCount ? sampleCount - (baseIndex + _tableIndexIncrement) : baseIndex + _tableIndexIncrement;
-	float nextPercentage = _currentTableIndex - baseIndex;
-	float basePercentage = 1.0f - nextPercentage;
-	float sample1 = tableWeight1 * (basePercentage * _tables[tableIndex1]->samples[baseIndex] + nextPercentage * _tables[tableIndex1]->samples[nextIndex]);
-	float sample2 = tableWeight2 * (basePercentage * _tables[tableIndex2]->samples[baseIndex] + nextPercentage * _tables[tableIndex2]->samples[nextIndex]);
-	float finalSample = sample1 + sample2;
+	double nextPercentage = _currentTableIndex - baseIndex;
+	double basePercentage = 1.0 - nextPercentage;
+	double sample1 = tableWeight1 * (basePercentage * _tables[tableIndex1]->samples[baseIndex] + nextPercentage * _tables[tableIndex1]->samples[nextIndex]);
+	double sample2 = tableWeight2 * (basePercentage * _tables[tableIndex2]->samples[baseIndex] + nextPercentage * _tables[tableIndex2]->samples[nextIndex]);
+	double finalSample = sample1 + sample2;
 
 	/* Increment and check index */
 	_currentTableIndex += _tableIndexIncrement;
 	if (_currentTableIndex >= sampleCount) {
-		_currentTableIndex = _currentTableIndex - (float)sampleCount;
+		_currentTableIndex = _currentTableIndex - (double)sampleCount;
 	}
 
 	return finalSample;
 }
 
-void Oscillator::setFrequencyValue(float value)
+void Oscillator::setFrequencyValue(double value)
 {
 	if (value > 1.0f)
 		value = 1.0f;
@@ -68,7 +68,7 @@ void Oscillator::setFrequencyValue(float value)
 	_tableIndexIncrement = _frequencyValue / convertToFrequencyValue(OSCILLATOR_BASE_FREQUENCY);
 }
 
-float Oscillator::getFrequencyValue()
+double Oscillator::getFrequencyValue()
 {
 	return _frequencyValue;
 }
