+++ Util.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-inline float convertRanges(float value, float oldMin, float oldMax, float newMin, float newMax)
+inline double convertRanges(double value, double oldMin, double oldMax, double newMin, double newMax)
 {
 	return (((value - oldMin) * (newMax - newMin)) / (oldMax - oldMin)) + newMin;
 }
